Prueba/Persona/src/main/java/com/micro1/Persona/service/PersonaService.java
package com.micro1.Persona.service;


import com.micro1.Persona.entity.Persona;
import com.micro1.Persona.repository.PersonaRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class PersonaService {
    @Autowired
    private PersonaRepository personaRepository;

    public Persona savePersona(Persona persona) {
        return personaRepository.save(persona);
    }

    public List<Persona> getAllPersonas() {
        return personaRepository.findAll();
    }
}
